new useful aliases
Conflicts:
	.bash_aliases

+++ .bash_aliases
@@ -1,10 +1,6 @@
 alias sudo=/usr/bin/sudo
-alias myjobs='sg_list -srp -usadali'
-alias monjobs='watch sg_list -srp -usadali'
-alias myjobs='sg_list -srp -usadali'
-alias monjobs='watch sg_list -srp -usadali'
-alias myjobs='sg_list -srp -usadali'
-alias monjobs='watch sg_list -srp -usadali'
+alias vncconfig='vncconfig -iconic'
+alias mi='micro'
 # Some usable functions
 export SCRIPTS="$RAID/local/buetext"
 function execute () {
@@ -14,13 +10,17 @@ function execute () {
 alias execute="$PYTHON  $SCRIPTS/"
 alias executewhat='execute what'
 
-function exp() {
-  cd /d4m/ears/expts/$1/sequences
-}
-function vid_exp() {
-  cd /d4m/vid/expts/$1/sequences
-}
-function bkp() {
- cp $1 $1.bak
+function pr() {
+ pushd /home/sadali/pr/$1
+ cd $(pwd -P)
 }
 
+function bkp() {
+ if [ ! -e $1.bak ]; then
+ cp $1 $1.bak
+ else
+ cp $1 $1.date
+ fi
+}
+
+source ~/.aliases
